# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "HST API Documentation",
        "description": "Application API services documentation page",
        "contact": {
            "email": "volkangurgor@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/campuses": {
            "get": {
                "tags": [
                    "Campuses"
                ],
                "summary": "Get all campuses",
                "operationId": "c23017978111f714dd601fd1268a7a5d",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Campuses"
                ],
                "summary": "Create a new campus",
                "operationId": "ab65a35f48b694f649bfcc7c41f6b9ba",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Campus A"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "Active"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/campuses/{id}": {
            "get": {
                "tags": [
                    "Campuses"
                ],
                "summary": "Get a specific campus",
                "operationId": "82b3fcb4e99a80c55f03f371f6aeef7a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Campus ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Campus not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Campuses"
                ],
                "summary": "Update a campus",
                "operationId": "0a0909ab64b3db580ee8f624269d73d5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Campus ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Updated Campus A"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "Inactive"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Campus not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Campuses"
                ],
                "summary": "Delete a campus",
                "operationId": "ed33826927d31021b94cba168d702bd0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Campus ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "404": {
                        "description": "Campus not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/branches": {
            "get": {
                "tags": [
                    "Branches"
                ],
                "summary": "Get all branches",
                "operationId": "47ae5f5d8dfc5541a89c6162a180ad51",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Branches"
                ],
                "summary": "Create a new branch",
                "operationId": "4d35e38c04746630efabc9acbed7c538",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Branch A"
                                    },
                                    "type": {
                                        "type": "string",
                                        "example": "Type A"
                                    },
                                    "campus_id": {
                                        "type": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/branches/{id}": {
            "get": {
                "tags": [
                    "Branches"
                ],
                "summary": "Get a specific branch",
                "operationId": "c84a98c56e97c7b78ef9cfaffd266ac1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Branch ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Branch not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Branches"
                ],
                "summary": "Update a branch",
                "operationId": "dcdb0ab40723f76e81ceedf8e326d3fc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Branch ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Updated Branch A"
                                    },
                                    "type": {
                                        "type": "string",
                                        "example": "Updated Type A"
                                    },
                                    "campus_id": {
                                        "type": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Branch not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Branches"
                ],
                "summary": "Delete a branch",
                "operationId": "37b47536a611d18a266f74fbb4505018",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Branch ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "404": {
                        "description": "Branch not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "in": "header",
                "name": "Authorization",
                "description": "Bearer token to access these API endpoints",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}